Client is a separate component now
